moveTo updated

+++ src/hillbillies/model/UnitTest.java
@@ -17,7 +17,7 @@ public class UnitTest {
 		assertTrue(Kobbe.getStrength()== 50);
 		assertTrue(Kobbe.getOrientation() == Math.PI / 2.0);
 		System.out.println(Arrays.toString(Kobbe.getPosition()));
-		double[] target = new double[]{1,2,3};
+		int[] target = new int[]{1,2,3};
 		Kobbe.moveTo(target);
 		Kobbe.advanceTime(0.1);
 		Kobbe.advanceTime(0.1);
@@ -58,7 +58,7 @@ public class UnitTest {
 		Kobbe.advanceTime(0.1);
 		Kobbe.advanceTime(0.1);
 		Kobbe.advanceTime(0.1);
-		Kobbe.moveTo(new double [] {0,0,0});
+		Kobbe.moveTo(new int[] {0,0,0});
 		Kobbe.advanceTime(0.1);
 		Kobbe.advanceTime(0.1);
 		Kobbe.advanceTime(0.1);
